Server/requirements.txt
pyOpenSSL==24.2.1
aiohttp==3.10.5
aiohttp_cors==0.7.0
aiohttp_index==0.1
aiodns==3.2.0
brotli==1.1.0
pywin32==306
pythonnet==3.0.3
elevate==0.1.3
watchdog==5.0.2
pyinstaller==6.10.0
wxpython==4.2.2
markdown==3.7
beautifulsoup4==4.12.3
